fake_news_analysis/app/requirements.txt
torch==1.12.0
transformers==4.15.0
psycopg2==2.9.9
scikit_learn==1.4.1.post1
python-decouple